Adding wielded treasure tables for generators
