+++ Source/ACE.Server/WorldObjects/Creature_Equipment.cs
@@ -415,65 +415,10 @@ namespace ACE.Server.WorldObjects
 
             var table = new TreasureWieldedTable(WieldedTreasure);
 
-            foreach (var set in table.Sets)
-                GenerateWieldedTreasureSet(set);
-        }
+            var wieldedTreasure = GenerateWieldedTreasureSets(table);
 
-        public void GenerateWieldedTreasureSet(TreasureWieldedSet set)
-        {
-            var rng = ThreadSafeRandom.Next(0.0f, set.TotalProbability);
-            var probability = 0.0f;
-
-            foreach (var item in set.Items)
-            {
-                probability += item.Item.Probability;
-                if (rng > probability) continue;
-
-                // item roll successful, spawn item in creature inventory
-                var success = CreateWieldedTreasure(item.Item);
-                if (!success) continue;
-
-                // traverse into possible subsets
-                if (item.Subset != null)
-                    GenerateWieldedTreasureSet(item.Subset);
-
-                break;
-            }
-        }
-
-        public bool CreateWieldedTreasure(TreasureWielded item)
-        {
-            var wo = WorldObjectFactory.CreateNewWorldObject(item.WeenieClassId);
-            if (wo == null) return false;
-
-            if (item.PaletteId > 0)
-                wo.PaletteTemplate = (int)item.PaletteId;
-
-            if (item.Shade > 0)
-                wo.Shade = item.Shade;
-
-            if (item.StackSize > 0)
-            {
-                // fix lugians only having 1 rock?
-                if (wo.Name.Equals("Rock") && item.StackSize == 1 && item.StackSizeVariance == 0)
-                {
-                    item.StackSize = 10;
-                    item.StackSizeVariance = 0.1f;
-                }
-
-                var stackSize = item.StackSize;
-
-                var hasVariance = item.StackSizeVariance > 0;
-                if (hasVariance)
-                {
-                    var minStack = (int)Math.Round(item.StackSize * item.StackSizeVariance);
-                    var maxStack = item.StackSize;
-                    stackSize = ThreadSafeRandom.Next(minStack, maxStack);
-                }
-                wo.StackSize = (ushort)stackSize;
-            }
-
-            return TryAddToInventory(wo);
+            foreach (var item in wieldedTreasure)
+                TryAddToInventory(item);
         }
     }
 }
